+++ DCM Framework/DCMPixelDataAttributeJPEG8.mm
@@ -182,7 +182,7 @@ fill_input_buffer (j_decompress_ptr cinfo)
 	
   data8_src_ptr src = (data8_src_ptr) cinfo->src;
   size_t nbytes;
-	nbytes = readFromData8(src->data, src->buffer, src->currentPosition, INPUT_BUF_SIZE);
+	nbytes = readFromData8(src->data, src->buffer, (int)src->currentPosition, INPUT_BUF_SIZE);
   //nbytes = JFREAD(src->infile, src->buffer, INPUT_BUF_SIZE);
 
   if (nbytes <= 0) {
