+++ DCM Framework/jpegdatasrc.mm
@@ -1,5 +1,5 @@
 //
-//  ©Alex Bettarini -- all rights reserved
+//  ï¿½Alex Bettarini -- all rights reserved
 //  License GPLv3.0 -- see License File
 //
 //  At the end of 2014 the project was forked from OsiriX to become Miele-LXIV
@@ -64,7 +64,7 @@ readFromData(NSData *data, JOCTET *buffer, int currentPosition, int length)
 		[data getBytes:buffer range:range];
     
 	//NSLog(@"LengthToRead %d", lengthToRead);
-	return lengthToRead;
+	return (int)lengthToRead;
 }
 
 
@@ -138,7 +138,7 @@ fill_input_buffer (j_decompress_ptr cinfo)
 {	
   data_src_ptr src = (data_src_ptr) cinfo->src;
   size_t nbytes;
-  nbytes = readFromData(src->data, src->buffer, src->currentPosition, INPUT_BUF_SIZE);
+  nbytes = readFromData(src->data, src->buffer, (int)src->currentPosition, INPUT_BUF_SIZE);
   //nbytes = JFREAD(src->infile, src->buffer, INPUT_BUF_SIZE);
 
   if (nbytes <= 0) {
